Add basic dedupe-service
This application is a very crude example of a deduplication service.
NB! Not meant for production!

The service excepts a request on the endpoint "verify" and expects atleast one query parameter (and identifier of some sorts). This identifier is stored for future lookups to verify if this identifier has already been processed

TRACE-111